feat(kfx): the kfx view contract as a shared package
@kungfu-tech/kfx carries the types a view extension implements and the
UI tokens it renders with — types erase and tokens are plain objects,
so the package bundles into each extension while react and the
capability SDK stay shell-injected. The workspace gains the
extensions/*/* glob so suite members nested under a suite directory
are workspace members.
